Set compile Warning level 4 in SVObserver Issue SVO-2798

+++ Operators/SVDLLToolLoadLibraryClass.cpp
@@ -65,9 +65,9 @@ HRESULT  SVDLLToolLoadLibraryClass::InitInputValueDef(SvStl::MessageContainer& r
 		{
 			Result = -12350;
 			if (!m_pfnGetInputValueDefinitions)              
-				rE.addMessage(SvStl::MessageData(-12359, SvStl::Tid_GetProcAddressError_GetInputValueDefinitions), false);
+				rE.addMessage(SvStl::MessageData(static_cast<DWORD> (-12359), SvStl::Tid_GetProcAddressError_GetInputValueDefinitions), false);
 			if (!m_pfnDestroyInputValueDefinitionStructures)  
-				rE.addMessage(SvStl::MessageData(-12360, SvStl::Tid_GetProcAddressError_DestroyInputValueDefinitionStructures), false);
+				rE.addMessage(SvStl::MessageData(static_cast<DWORD> (-12360), SvStl::Tid_GetProcAddressError_DestroyInputValueDefinitionStructures), false);
 
 
 		}
@@ -108,11 +108,11 @@ HRESULT  SVDLLToolLoadLibraryClass::InitResultValueDef(SvStl::MessageContainer& 
 			rE.setMessage(static_cast<DWORD> (Result), SvStl::Tid_Empty, SvStl::SourceFileParams(StdMessageParams));
 			if (!m_pfnDestroyInputValueDefinitionStructures)
 			{
-				rE.addMessage(SvStl::MessageData(-12360, SvStl::Tid_GetProcAddressError_DestroyInputValueDefinitionStructures), false);
+				rE.addMessage(SvStl::MessageData(static_cast<DWORD> (-12360), SvStl::Tid_GetProcAddressError_DestroyInputValueDefinitionStructures), false);
 			}
 			if (!m_pfnDestroyResultValueDefinitionStructures)
 			{
-				rE.addMessage(SvStl::MessageData(-12361, SvStl::Tid_GetProcAddressError_DestroyResultValueDefinitionStructures), false);
+				rE.addMessage(SvStl::MessageData(static_cast<DWORD> (-12361), SvStl::Tid_GetProcAddressError_DestroyResultValueDefinitionStructures), false);
 			}
 		}
 	}
@@ -256,21 +256,21 @@ HRESULT SVDLLToolLoadLibraryClass::Open(LPCTSTR p_szLibrary, SVDllLoadLibraryCal
 				if (!m_pfnSetMILResultImages && !m_pfnSetMILInputImages
 					&& !m_pfnSetHBITMAPInputImages && !m_pfnGetHBITMAPResultImages)
 				{
-					e.addMessage(SvStl::MessageData(-12373, SvStl::Tid_ToolLoadError_NoImageDef), false);
+					e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12373), SvStl::Tid_ToolLoadError_NoImageDef), false);
 				}
 				else if (m_pfnSetMILResultImages && m_pfnSetMILInputImages
 					&&  m_pfnSetHBITMAPInputImages && m_pfnGetHBITMAPResultImages)
 				{
-					e.addMessage(SvStl::MessageData(-12375, SvStl::Tid_ToolLoadError_BothSetsImagesDef), false);
+					e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12375), SvStl::Tid_ToolLoadError_BothSetsImagesDef), false);
 				}
 				else if ((m_pfnSetMILResultImages && (m_pfnGetHBITMAPResultImages || m_pfnSetHBITMAPInputImages))
 					|| (m_pfnSetHBITMAPInputImages && (m_pfnSetMILInputImages || m_pfnSetMILResultImages)))
 				{
-					e.addMessage(SvStl::MessageData(-12376, SvStl::Tid_ToolLoadError_BothImagesDef), false);
+					e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12376), SvStl::Tid_ToolLoadError_BothImagesDef), false);
 				}
 				else
 				{
-					e.addMessage(SvStl::MessageData(-12374, SvStl::Tid_ToolLoadError_InFunctionDef), false);
+					e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12374), SvStl::Tid_ToolLoadError_InFunctionDef), false);
 				}
 
 			}// end else	// error in MIL/HBITMAP definitions
@@ -370,22 +370,22 @@ HRESULT SVDLLToolLoadLibraryClass::Open(LPCTSTR p_szLibrary, SVDllLoadLibraryCal
 			Result = -12350;
 			e.setMessage(static_cast<DWORD> (Result), SvStl::Tid_Empty, SvStl::SourceFileParams(StdMessageParams));
 
-			if (!m_pfnSimpleTest)                             e.addMessage(SvStl::MessageData(-12351, SvStl::Tid_GetProcAddressError_SimpleTest), false);
-			if (!m_pfnGetToolName)                            e.addMessage(SvStl::MessageData(-12352, SvStl::Tid_GetProcAddressError_GetToolName), false);
-			if (!m_pfnGetToolVersion)                         e.addMessage(SvStl::MessageData(-12353, SvStl::Tid_GetProcAddressError_GetToolVersion), false);
-			if (!m_pfnRunTool)                                e.addMessage(SvStl::MessageData(-12354, SvStl::Tid_GetProcAddressError_RunTool), false);
-			if (!m_pfnStartup)                                e.addMessage(SvStl::MessageData(-12355, SvStl::Tid_GetProcAddressError_Startup), false);
-			if (!m_pfnShutDown)                               e.addMessage(SvStl::MessageData(-12356, SvStl::Tid_GetProcAddressError_ShutDown), false);
-			if (!m_pfnInitializeRun)                          e.addMessage(SvStl::MessageData(-12357, SvStl::Tid_GetProcAddressError_InitializeRun), false);
-			if (!m_pfnUninitializeRun)                        e.addMessage(SvStl::MessageData(-12358, SvStl::Tid_GetProcAddressError_UninitializeRun), false);
-			if (!m_pfnSetInputValues)                         e.addMessage(SvStl::MessageData(-12362, SvStl::Tid_GetProcAddressError_SetInputValues), false);
-			if (!m_pfnGetResultValues)                        e.addMessage(SvStl::MessageData(-12363, SvStl::Tid_GetProcAddressError_GetResultValues), false);
-			if (!m_pfnGetMessageString)                       e.addMessage(SvStl::MessageData(-12364, SvStl::Tid_GetProcAddressError_GetMessageString), false);
-			if (!m_pfnValidateValueParameter)                 e.addMessage(SvStl::MessageData(-12365, SvStl::Tid_GetProcAddressError_ValidateValueParameter), false);
-			if (!m_pfnGetResultValueDefinitions)              e.addMessage(SvStl::MessageData(-12366, SvStl::Tid_GetProcAddressError_GetResultValueDefinitions), false);
-			if (!m_pfnGetNumberOfInputImages)                 e.addMessage(SvStl::MessageData(-12367, SvStl::Tid_GetProcAddressError_GetNumberOfInputImages), false);
-			if (!m_pfnGetResultImageDefinitions)              e.addMessage(SvStl::MessageData(-12368, SvStl::Tid_GetProcAddressError_GetResultImageDefinitions), false);
-			if (!m_pfnDestroyImageDefinitionStructure)        e.addMessage(SvStl::MessageData(-12369, SvStl::Tid_GetProcAddressError_DestroyImageDefinitionStructure), false);
+			if (!m_pfnSimpleTest)                      e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12351), SvStl::Tid_GetProcAddressError_SimpleTest), false);
+			if (!m_pfnGetToolName)                     e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12352), SvStl::Tid_GetProcAddressError_GetToolName), false);
+			if (!m_pfnGetToolVersion)                  e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12353), SvStl::Tid_GetProcAddressError_GetToolVersion), false);
+			if (!m_pfnRunTool)                         e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12354), SvStl::Tid_GetProcAddressError_RunTool), false);
+			if (!m_pfnStartup)                         e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12355), SvStl::Tid_GetProcAddressError_Startup), false);
+			if (!m_pfnShutDown)                        e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12356), SvStl::Tid_GetProcAddressError_ShutDown), false);
+			if (!m_pfnInitializeRun)                   e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12357), SvStl::Tid_GetProcAddressError_InitializeRun), false);
+			if (!m_pfnUninitializeRun)                 e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12358), SvStl::Tid_GetProcAddressError_UninitializeRun), false);
+			if (!m_pfnSetInputValues)                  e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12362), SvStl::Tid_GetProcAddressError_SetInputValues), false);
+			if (!m_pfnGetResultValues)                 e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12363), SvStl::Tid_GetProcAddressError_GetResultValues), false);
+			if (!m_pfnGetMessageString)                e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12364), SvStl::Tid_GetProcAddressError_GetMessageString), false);
+			if (!m_pfnValidateValueParameter)          e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12365), SvStl::Tid_GetProcAddressError_ValidateValueParameter), false);
+			if (!m_pfnGetResultValueDefinitions)       e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12366), SvStl::Tid_GetProcAddressError_GetResultValueDefinitions), false);
+			if (!m_pfnGetNumberOfInputImages)          e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12367), SvStl::Tid_GetProcAddressError_GetNumberOfInputImages), false);
+			if (!m_pfnGetResultImageDefinitions)       e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12368), SvStl::Tid_GetProcAddressError_GetResultImageDefinitions), false);
+			if (!m_pfnDestroyImageDefinitionStructure) e.addMessage(SvStl::MessageData(static_cast<DWORD> (-12369), SvStl::Tid_GetProcAddressError_DestroyImageDefinitionStructure), false);
 		}
 	}// end if ( nullptr != m_hmHandle )
